requirements.txt
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
kaggle==1.7.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==77.0.3
six==1.17.0
sklearn-compat==0.1.3
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
